+++ CHANGELOG.md
@@ -9,10 +9,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 ### Added
 
 1. The table that is written to within QuestDB can now be specified with `--questdb-table`.
+1. A timeout for QuestDB connections can now be specified with `--questdb-timeout`.
 
 ### Changed
 
-1. QuestDB connections time out after 10 seconds with an appropriate error message.
+1. QuestDB connections by default time out after 10 seconds with an appropriate error message.
 
 ## [1.2.4] - 2024-04-19